#13 - Move Color class into Console/ directory
