+++ src/Console/Color.php
@@ -10,7 +10,7 @@ declare(strict_types=1);
  * the LICENSE file that was distributed with this source code.
  */
 
-namespace Phalcon\Migrations\Script;
+namespace Phalcon\Migrations\Console;
 
 /**
  * Allows to generate messages using colors on xterm, ddterm, linux, etc.
@@ -107,9 +107,9 @@ final class Color
     /**
      * Identify if console supports colors
      *
-     * @return boolean
+     * @return bool
      */
-    public static function isSupportedShell()
+    public static function isSupportedShell(): bool
     {
         if (strtoupper(substr(PHP_OS, 0, 3)) === 'WIN') {
             return false !== getenv('ANSICON') || 'ON' === getenv('ConEmuANSI') || 'xterm' === getenv('TERM');
@@ -129,7 +129,7 @@ final class Color
      *
      * @return string
      */
-    public static function colorize($string, $fg = null, $at = null, $bg = null)
+    public static function colorize(string $string, int $fg = null, int $at = null, int $bg = null): string
     {
         // Shell not supported, exit early
         if (!static::isSupportedShell()) {
@@ -159,7 +159,11 @@ final class Color
         return $colored;
     }
 
-    public static function head($msg)
+    /**
+     * @param string $msg
+     * @return string
+     */
+    public static function head(string $msg): string
     {
         return static::colorize($msg, Color::FG_BROWN);
     }
@@ -171,10 +175,10 @@ final class Color
      * @param string $msg
      * @return string
      */
-    public static function error($msg)
+    public static function error(string $msg): string
     {
         $msg = 'Error: ' . $msg;
-        $space = strlen($msg) + 4;
+        $space = self::tabSpaces($msg);
         $out  = static::colorize(str_pad(' ', $space), Color::FG_WHITE, Color::AT_BOLD, Color::BG_RED) . PHP_EOL;
         $out .= static::colorize('  ' . $msg . '  ', Color::FG_WHITE, Color::AT_BOLD, Color::BG_RED) . PHP_EOL;
         $out .= static::colorize(str_pad(' ', $space), Color::FG_WHITE, Color::AT_BOLD, Color::BG_RED) . PHP_EOL;
@@ -189,10 +193,10 @@ final class Color
      * @param string $msg
      * @return string
      */
-    public static function success($msg)
+    public static function success(string $msg): string
     {
         $msg = 'Success: ' . $msg;
-        $space = strlen($msg) + 4;
+        $space = self::tabSpaces($msg);
         $out  = static::colorize(str_pad(' ', $space), Color::FG_WHITE, Color::AT_BOLD, Color::BG_GREEN) . PHP_EOL;
         $out .= static::colorize('  ' . $msg . '  ', Color::FG_WHITE, Color::AT_BOLD, Color::BG_GREEN) . PHP_EOL;
         $out .= static::colorize(str_pad(' ', $space), Color::FG_WHITE, Color::AT_BOLD, Color::BG_GREEN) . PHP_EOL;
@@ -207,14 +211,28 @@ final class Color
      * @param string $msg
      * @return string
      */
-    public static function info($msg)
+    public static function info(string $msg): string
     {
         $msg = 'Info: ' . $msg;
-        $space = strlen($msg) + 4;
+        $space = self::tabSpaces($msg);
         $out  = static::colorize(str_pad(' ', $space), Color::FG_WHITE, Color::AT_BOLD, Color::BG_BLUE) . PHP_EOL;
         $out .= static::colorize('  ' . $msg . '  ', Color::FG_WHITE, Color::AT_BOLD, Color::BG_BLUE) . PHP_EOL;
         $out .= static::colorize(str_pad(' ', $space), Color::FG_WHITE, Color::AT_BOLD, Color::BG_BLUE) . PHP_EOL;
 
         return $out;
+    }
+
+    /**
+     * Output tab space
+     *
+     * Depending on length of string.
+     *
+     * @param string $string
+     * @param int $tabSize
+     * @return int
+     */
+    protected static function tabSpaces(string $string, int $tabSize = 4): int
+    {
+        return strlen($string) + $tabSize;
     }
 }
