+++ src/Migrations.php
@@ -20,13 +20,13 @@ use Phalcon\Db\Adapter\AdapterInterface;
 use Phalcon\Db\Column;
 use Phalcon\Db\Exception as DbException;
 use Phalcon\Db\Index;
+use Phalcon\Migrations\Console\Color;
 use Phalcon\Migrations\Console\OptionStack;
 use Phalcon\Migrations\Db\Dialect\DialectMysql;
 use Phalcon\Migrations\Db\Dialect\DialectPostgresql;
 use Phalcon\Migrations\Mvc\Model\Migration as ModelMigration;
 use Phalcon\Migrations\Mvc\Model\Migration\TableAware\ListTablesDb;
 use Phalcon\Migrations\Mvc\Model\Migration\TableAware\ListTablesIterator;
-use Phalcon\Migrations\Script\Color;
 use Phalcon\Migrations\Script\ScriptException;
 use Phalcon\Migrations\Version\IncrementalItem;
 use Phalcon\Migrations\Version\ItemCollection as VersionCollection;
